reorganize vm structure
