+++ src/hardware/vm.rs
@@ -1,5 +1,7 @@
 use crate::hardware::{memory::Memory, registers::Registers};
 use crate::instructions::*;
+use std::{env, fs::File, io::{BufReader, Read}};
+use byteorder::{BigEndian, ReadBytesExt};
 
 const PC_START: u16 = 0x3000;
 
@@ -24,16 +26,30 @@ impl VM {
         }
     }
 
+    pub fn load_arguments(&mut self) {
+        let args: Vec<String> = env::args().collect();
+    
+        if args.len() < 2 {
+            eprintln!("lc3 [image-file1] ...");
+            return;
+        }
+    
+        // Iterate over each argument (skipping the first one which is the program name)
+        for arg in &args[1..] {
+            self.read_image_file(arg);
+        }
+    }
+
     pub fn run(&mut self) {
         self.reg.pc = PC_START;
         while self.running {
             // Fetch instruction from memory
             let instruction_address: usize = self.reg.pc.into();
             let instruction = self.mem.read(instruction_address);
-
+            
             // Increment PC
             self.reg.pc += 1;
-
+            
             // Decode opcode
             let raw_opcode = instruction >> 12;
             match Opcode::try_from(raw_opcode) {
@@ -46,7 +62,7 @@ impl VM {
         }
         // Shutdown goes here.
     }
-
+    
     fn execute_instruction(&mut self, opcode: Opcode, instr: u16) {
         match opcode {
             Opcode::OpAdd => self.op_add(instr),
@@ -66,6 +82,19 @@ impl VM {
             Opcode::OpRes => {}
             Opcode::OpRti => {}
             // The last 2 are unused opcodes, I have to define what to do when they are called.
+        }
+    }
+
+    fn read_image_file(&mut self, file_path: &str) {
+        let file = File::open(file_path).unwrap();
+        let mut reader = BufReader::new(file);
+    
+        let mut address = reader
+            .read_u16::<BigEndian>()
+            .unwrap();
+        while let Ok(instr) = reader.read_u16::<BigEndian>() {
+            self.mem.write(address as usize, instr);
+            address += 1;
         }
     }
 }
