+++ src/lib.rs
@@ -1,3 +1,2 @@
 pub mod hardware;
 pub mod instructions;
-pub mod utils;
